+++ src/pamietacz/production.py
@@ -4,4 +4,4 @@ DEBUG = True
 TEMPLATE_DEBUG = DEBUG
 DEBUG_PROPAGATE_EXCEPTIONS = DEBUG
 
-MEDIA_ROOT = "unit_tests_uploaded/"
+MEDIA_ROOT = "uploaded/"